backend/requirements.txt
# FastAPI ve web framework - Python 3.13 uyumlu
fastapi
uvicorn[standard]
python-multipart

# Veritabanı - SQLite
sqlalchemy
aiosqlite

# Pydantic - en son sürüm (3.13 uyumlu)
pydantic>=2.10

# HTTP client
httpx

# Excel
openpyxl

# Yardımcı
python-dotenv
